pyproject.toml
[project]
name = "pyowl"
version = "0.1.0"
description = "This module provides high-level control for the [OWL platform](#)."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyserial>=3.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pyowl"]
